Modify graph data from outside graph module

+++ lib/lineGeometry.h
@@ -7,9 +7,9 @@ class LineGeometry : public QQuick3DGeometry
 {
     Q_OBJECT
     QML_NAMED_ELEMENT(LineGeometry)
-    Q_PROPERTY(float thickness MEMBER thickness_ NOTIFY dataChanged)
-    Q_PROPERTY(QList<float> xs MEMBER xs_ NOTIFY dataChanged)
-    Q_PROPERTY(QList<float> ys MEMBER ys_ NOTIFY dataChanged)
+    Q_PROPERTY(double thickness MEMBER thickness_ NOTIFY dataChanged)
+    Q_PROPERTY(QList<double> xs MEMBER xs_ NOTIFY dataChanged)
+    Q_PROPERTY(QList<double> ys MEMBER ys_ NOTIFY dataChanged)
 
 public:
     LineGeometry();
@@ -20,6 +20,6 @@ Q_SIGNALS:
 
 private:
     void updateData();
-    float thickness_ = 0.01;
-    QList<float> xs_, ys_;
+    double thickness_ = 0.01;
+    QList<double> xs_, ys_;
 };
